+++ circus-train-api/src/main/java/com/hotels/bdp/circustrain/api/copier/CompositeCopierFactory.java
@@ -82,21 +82,20 @@ public class CompositeCopierFactory implements CopierFactory {
   }
 
   @Override
-  public Copier newInstance(
-      String eventId,
-      Path sourceBaseLocation,
-      List<Path> sourceSubLocations,
-      Path replicaLocation,
-      Map<String, Object> copierOptions) {
+  public Copier newInstance(CopierContext copierContext) {
     List<Copier> copiers = new ArrayList<>(delegates.size());
     int i = 0;
     for (CopierFactory delegate : delegates) {
-      CopierPathGeneratorParams copierPathGeneratorParams = CopierPathGeneratorParams.newParams(i++, eventId,
-          sourceBaseLocation, sourceSubLocations, replicaLocation, copierOptions);
+      CopierPathGeneratorParams copierPathGeneratorParams = CopierPathGeneratorParams
+          .newParams(i++, copierContext.getEventId(), copierContext.getSourceBaseLocation(),
+              copierContext.getSourceSubLocations(), copierContext.getReplicaLocation(),
+              copierContext.getCopierOptions());
       Path newSourceBaseLocation = pathGenerator.generateSourceBaseLocation(copierPathGeneratorParams);
       Path newReplicaLocation = pathGenerator.generateReplicaLocation(copierPathGeneratorParams);
-      Copier copier = delegate.newInstance(eventId, newSourceBaseLocation, sourceSubLocations, newReplicaLocation,
-          copierOptions);
+
+      CopierContext delegateContext = new CopierContext(copierContext.getEventId(), newSourceBaseLocation,
+          copierContext.getSourceSubLocations(), newReplicaLocation, copierContext.getCopierOptions());
+      Copier copier = delegate.newInstance(delegateContext);
       copiers.add(copier);
     }
     return new CompositeCopier(copiers, metricsMerger);
@@ -108,17 +107,20 @@ public class CompositeCopierFactory implements CopierFactory {
       Path sourceBaseLocation,
       Path replicaLocation,
       Map<String, Object> copierOptions) {
-    List<Copier> copiers = new ArrayList<>(delegates.size());
-    int i = 0;
-    for (CopierFactory delegatee : delegates) {
-      CopierPathGeneratorParams copierPathGeneratorParams = CopierPathGeneratorParams.newParams(i++, eventId,
-          sourceBaseLocation, null, replicaLocation, copierOptions);
-      Path newReplicaLocation = pathGenerator.generateReplicaLocation(copierPathGeneratorParams);
-      Path newSourceBaseLocation = pathGenerator.generateSourceBaseLocation(copierPathGeneratorParams);
-      Copier copier = delegatee.newInstance(eventId, newSourceBaseLocation, newReplicaLocation, copierOptions);
-      copiers.add(copier);
-    }
-    return new CompositeCopier(copiers, metricsMerger);
+    CopierContext copierContext = new CopierContext(eventId, sourceBaseLocation, replicaLocation, copierOptions);
+    return newInstance(copierContext);
+  }
+
+  @Override
+  public Copier newInstance(
+      String eventId,
+      Path sourceBaseLocation,
+      List<Path> sourceSubLocations,
+      Path replicaLocation,
+      Map<String, Object> copierOptions) {
+    CopierContext copierContext = new CopierContext(eventId, sourceBaseLocation, sourceSubLocations, replicaLocation,
+        copierOptions);
+    return newInstance(copierContext);
   }
 
 }