+++ circus-train-avro/src/test/java/com/hotels/bdp/circustrain/avro/util/SchemaCopierTest.java
@@ -17,6 +17,7 @@ package com.hotels.bdp.circustrain.avro.util;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
+import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.when;
 
@@ -36,6 +37,7 @@ import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import com.hotels.bdp.circustrain.api.copier.Copier;
+import com.hotels.bdp.circustrain.api.copier.CopierContext;
 import com.hotels.bdp.circustrain.api.copier.CopierFactory;
 import com.hotels.bdp.circustrain.api.copier.CopierFactoryManager;
 import com.hotels.bdp.circustrain.api.copier.CopierOptions;
@@ -70,7 +72,7 @@ public class SchemaCopierTest {
     copierOptionsMap.put(CopierOptions.COPY_DESTINATION_IS_FILE, "true");
     when(copierFactoryManager.getCopierFactory(eq(source), eq(targetFile), eq(copierOptionsMap)))
         .thenReturn(copierFactory);
-    when(copierFactory.newInstance(eq(eventId), eq(source), eq(targetFile), eq(copierOptionsMap))).thenReturn(copier);
+    when(copierFactory.newInstance(any(CopierContext.class))).thenReturn(copier);
     when(copier.copy()).thenReturn(metrics);
     when(metrics.getBytesReplicated()).thenReturn(123L);
     Path result = schemaCopier.copy(source.toString(), destination.toString(), eventTableReplication, eventId);